+++ client/src/components/application/dates/completenessDates.ts
@@ -1,4 +1,4 @@
-import { EndDate, StartDate } from "../applicationDates";
+import { EndDate, StartDate } from "./applicationDates";
 
 /**
  * Completeness Phase
