+++ client/src/components/application/dates/index.ts
@@ -1,4 +1,4 @@
 export * from "./applicationDates";
-export * from "./concept/conceptDates";
-export * from "./state-application/stateApplicationDates";
-export * from "./completeness/completenessDates";
+export * from "./conceptDates";
+export * from "./stateApplicationDates";
+export * from "./completenessDates";
